Add PC port infrastructure using PsyCross (SDL2 + OpenGL)
Port the Silent Hill PS1 decompilation to PC using PsyCross as a PSX
hardware abstraction layer. The game boots through all startup screens
(Boot -> Konami -> KCET -> MovieIntro -> MainMenu) with visual output
rendering to an SDL window.

Key components:
- CMake build system targeting MSYS2/MinGW64
- PSY-Q header shims bridging to PsyCross equivalents
- PSX memory emulation (2MB RAM, scratchpad, address translation)
- GS library stubs implementing GsInitGraph, GsDrawOt, etc.
- Sound/pad/API stubs for unimplemented subsystems
- Map overlay stubs (statically linked instead of dynamic loading)

Rendering fixes:
- Disable PGXP (VERTTYPE half/short mismatch causes NaN vertices)
- Force dfe=1 in PsyCross (PSX dfe flag misused as render target selector)
- Eliminate PSX VRAM double-buffer offsets (draw env offset pushed content off-screen)
- Center coordinate system (set draw offset to screen_w/2, screen_h/2)
- Fix P_TAG/setCodeWord for 64-bit (12-byte P_TAG vs PSX 4-byte)

Runtime fixes for 64-bit compatibility:
- PSX_ADDR macro for translating PSX addresses to emulated RAM
- uintptr_t casts replacing s32 pointer truncation
- Screen fade acceleration (4x speed on PC)
- Skip hardware-dependent init (vibration, memory cards)
- Delta time injection in internal loops

+++ src/bodyprog/gfx/bodyprog_80040A64.c
@@ -1564,18 +1564,18 @@ void IpdHeader_FixHeaderOffsets(s_IpdHeader* ipdHdr) // 0x80043DA4
 {
     s_IpdModelBuffer* curModelBuf;
 
-    ipdHdr->lmHdr_4           = (u8*)ipdHdr->lmHdr_4 + (u32)ipdHdr;
-    ipdHdr->modelInfo_14      = (u8*)ipdHdr->modelInfo_14 + (u32)ipdHdr;
-    ipdHdr->modelBuffers_18   = (u8*)ipdHdr->modelBuffers_18 + (u32)ipdHdr;
-    ipdHdr->modelOrderList_50 = (u8*)ipdHdr->modelOrderList_50 + (u32)ipdHdr;
+    ipdHdr->lmHdr_4           = (u8*)ipdHdr->lmHdr_4 + (uintptr_t)ipdHdr;
+    ipdHdr->modelInfo_14      = (u8*)ipdHdr->modelInfo_14 + (uintptr_t)ipdHdr;
+    ipdHdr->modelBuffers_18   = (u8*)ipdHdr->modelBuffers_18 + (uintptr_t)ipdHdr;
+    ipdHdr->modelOrderList_50 = (u8*)ipdHdr->modelOrderList_50 + (uintptr_t)ipdHdr;
 
     for (curModelBuf = &ipdHdr->modelBuffers_18[0];
          curModelBuf < &ipdHdr->modelBuffers_18[ipdHdr->modelBufferCount_9];
          curModelBuf++)
     {
-        curModelBuf->field_C  = (u8*)curModelBuf->field_C + (u32)ipdHdr;
-        curModelBuf->field_10 = (u8*)curModelBuf->field_10 + (u32)ipdHdr;
-        curModelBuf->field_14 = (u8*)curModelBuf->field_14 + (u32)ipdHdr;
+        curModelBuf->field_C  = (u8*)curModelBuf->field_C + (uintptr_t)ipdHdr;
+        curModelBuf->field_10 = (u8*)curModelBuf->field_10 + (uintptr_t)ipdHdr;
+        curModelBuf->field_14 = (u8*)curModelBuf->field_14 + (uintptr_t)ipdHdr;
     }
 }
